V2 add ProcessingPrepaymentBalanceInCents to AccountBalance
Add new add ProcessingPrepaymentBalanceInCents attribute to
AccountBalance. The ProcessingPrepaymentBalanceInCents attribute
is a similar format to the BalanceInCents attribute and contains
the total for processing prepayment credit invoices in each currency.
This value is useful when trying to determine if the customer's
prepayment balance has run out or if it is currently processing
while waiting on a payment for the corresponding purchase invoice.